+++ CHANGELOG.md
@@ -72,6 +72,9 @@
         - E-mails
         - HTML-objects
         - Interesting Responses
+- Reports -- If a directory has been passed as an ```outfile``` option the
+    report will be written under that directory using the default ```outfile```
+    value as a filename.
 - Issues
     - Added attribute ```remarks``` holding a ```Hash``` of remarks about
         that issue with the entity which made the remark as _key_ and an ```Array```